image-check/requirements.txt
Flask==2.2.2
requests==2.28.1
ptvsd==4.3.2 # Required for debugging.
gunicorn==20.1.0
google-api-core==2.8.2
google-auth==2.11.0
google-cloud-bigquery==2.34.4
google-cloud-pubsub==2.13.6
google-cloud-storage
pandas
google-cloud-secret-manager